script: Pass `&mut JSContext` to `navigate`
Also move the related `navigate_to_fragment` method
next to it.

Part of #40600
