+++ components/script/dom/location.rs
@@ -19,7 +19,6 @@ use crate::dom::globalscope::GlobalScope;
 use crate::dom::urlhelper::UrlHelper;
 use crate::dom::window::Window;
 use crate::navigation::navigate;
-use crate::script_runtime::CanGc;
 
 #[derive(PartialEq)]
 pub(crate) enum NavigationType {
@@ -84,13 +83,7 @@ impl Location {
             history_handling
         };
         // Step 4. Navigate navigable to url using sourceDocument, with exceptionsEnabled set to true and historyHandling set to historyHandling.
-        navigate(
-            navigable,
-            history_handling,
-            false,
-            load_data,
-            CanGc::from_cx(cx),
-        );
+        navigate(cx, navigable, history_handling, false, load_data);
     }
 
     /// Navigate the relevant `Document`'s browsing context.
@@ -163,11 +156,11 @@ impl Location {
             source_document.creation_sandboxing_flag_set_considering_parent_iframe(),
         );
         navigate(
+            cx,
             &self.window,
             history_handling,
             reload_triggered,
             load_data,
-            CanGc::from_cx(cx),
         );
     }
