+++ components/script/timers.rs
@@ -151,9 +151,7 @@ impl OneshotTimerCallback {
             OneshotTimerCallback::JsTimer(task) => task.invoke(this, js_timers, cx),
             #[cfg(feature = "testbinding")]
             OneshotTimerCallback::TestBindingCallback(callback) => callback.invoke(),
-            OneshotTimerCallback::RefreshRedirectDue(callback) => {
-                callback.invoke(CanGc::from_cx(cx))
-            },
+            OneshotTimerCallback::RefreshRedirectDue(callback) => callback.invoke(cx),
             OneshotTimerCallback::RunStepsAfterTimeout { completion, .. } => {
                 // <https://html.spec.whatwg.org/multipage/#run-steps-after-a-timeout>
                 // Step 4.4 Perform completionSteps.
